chore: Results페이지 폴더구조 수정
