+++ src/pages/Results/ChannelItem/ChannelItem.module.css
